Fixed issues with flashing. (#2)
* Added .img flashing to setup.sh

* Wording changes.

Since the rootfs installer can be used with both GSI and device-specific
images, changed wording to be more generic so it applies to both cases.

* Fixed extra imgs flashing.

The contents of /dev/block/platform seems to differ from device to
device and doesn't list partitions on any of the devices in my
knowledge. However, /dev/block/by-name exists on them all, so switched
to check partitions from there instead.
Fixed related dd commands too. dd doesn't like the use of * symbols so
using find to get the correct filepath.